+++ tests/CodeSnippetTest.php
@@ -171,4 +171,27 @@ class CodeSnippetTest extends TestCase
 
         $this->assertMatchesSnapshot((string)$snippet);
     }
+
+    /** @test */
+    public function it_returns_the_line_numbers()
+    {
+        $snippet = (new CodeSnippet())
+            ->surroundingLines(10, 12)
+            ->snippetLineCount(8)
+            ->fromFile($this->testsPath('data/file3.php'));
+
+        $this->assertCount(8, $snippet->getLineNumbers());
+        $this->assertMatchesSnapshot($snippet->getLineNumbers());
+    }
+
+    /** @test */
+    public function it_returns_the_selected_bounds()
+    {
+        $snippet = (new CodeSnippet())
+            ->surroundingLines(10, 12)
+            ->snippetLineCount(8)
+            ->fromFile($this->testsPath('data/file3.php'));
+
+        $this->assertEquals([10, 11, 12], range($snippet->getSelectedBounds()->start, $snippet->getSelectedBounds()->end));
+    }
 }